requirements.txt
flask
flask_cors
python_dotenv
pillow
requests
xmltodict